fix iw debugging manager reference in kernels.
