+++ src/interactive-window/serviceRegistry.web.ts
@@ -20,12 +20,13 @@ import {
 } from './editor-integration/types';
 import { InteractiveWindowCommandListener } from './interactiveWindowCommandListener';
 import { InteractiveWindowProvider } from './interactiveWindowProvider';
-import { IExportCommands, IInteractiveWindowProvider } from './types';
+import { IExportCommands, IInteractiveWindowDebuggingManager, IInteractiveWindowProvider } from './types';
 import { CodeGeneratorFactory } from './editor-integration/codeGeneratorFactory';
 import { GeneratedCodeStorageFactory } from './editor-integration/generatedCodeStorageFactory';
 import { IGeneratedCodeStorageFactory } from './editor-integration/types';
 import { GeneratedCodeStorageManager } from './generatedCodeStoreManager';
 import { InteractiveWindowTracebackFormatter } from './outputs/tracebackFormatter';
+import { InteractiveWindowDebuggingManager } from './debugger/jupyter/debuggingManager';
 
 export function registerTypes(serviceManager: IServiceManager) {
     serviceManager.addSingleton<IInteractiveWindowProvider>(IInteractiveWindowProvider, InteractiveWindowProvider);
@@ -54,4 +55,10 @@ export function registerTypes(serviceManager: IServiceManager) {
         GeneratedCodeStorageFactory
     );
     serviceManager.addSingleton<ITracebackFormatter>(ITracebackFormatter, InteractiveWindowTracebackFormatter);
+    serviceManager.addSingleton<IInteractiveWindowDebuggingManager>(
+        IInteractiveWindowDebuggingManager,
+        InteractiveWindowDebuggingManager,
+        undefined,
+        [IExtensionSingleActivationService]
+    );
 }
